requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0

# Statistics & ML
scipy>=1.10.0
statsmodels>=0.14.0
scikit-learn>=1.3.0

# HTTP & WebSocket
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.0

# Scheduling
apscheduler>=3.10.0

# Database
sqlalchemy>=2.0.0

# Logging
loguru>=0.7.0

# Telegram
python-telegram-bot>=20.0

# Config
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0
# Backtest & Visualization
yfinance>=0.2.0
matplotlib>=3.7.0

# Development
pytest>=7.0.0
